Refactor VoicesAdapter to use VoicesEditActions
Extracted `VoicesEditActions` and placed it in the repository package. `VoicesRepository` now implements `VoicesEditActions`.
This allows `VoicesAdapter` and `VoicesActions` to only rely on the narrower `VoicesEditActions` interface rather than the full `VoicesRepository`, cleanly reducing UI coupling without the boilerplate of proxying interfaces manually at call sites.
