+++ app/src/main/java/org/ole/planet/myplanet/ui/voices/ReplyActivity.kt
@@ -28,7 +28,6 @@ import org.ole.planet.myplanet.R
 import org.ole.planet.myplanet.callback.OnNewsItemClickListener
 import org.ole.planet.myplanet.databinding.ActivityReplyBinding
 import org.ole.planet.myplanet.model.News
-import org.ole.planet.myplanet.ui.voices.VoicesEditActions
 import org.ole.planet.myplanet.model.UserEntity
 import org.ole.planet.myplanet.repository.ActivitiesRepository
 import org.ole.planet.myplanet.repository.VoicesRepository
@@ -152,17 +151,7 @@ open class ReplyActivity : AppCompatActivity(), OnNewsItemClickListener {
                     },
                     onAnimateTyping = VoicesAdapterHelper.createOnAnimateTyping(lifecycleScope, dispatcherProvider),
                     labelManager = labelManager,
-                    voicesEditActions = object : VoicesEditActions {
-                        override suspend fun editPost(newsId: String, message: String, imagesToRemove: Set<String>, newImages: List<String>?): News? {
-                            return voicesRepository.editPost(newsId, message, imagesToRemove, newImages)
-                        }
-                        override suspend fun postReply(message: String, news: News, currentUser: UserEntity, imageList: List<String>?) {
-                            voicesRepository.postReply(message, news, currentUser, imageList)
-                        }
-                        override suspend fun getNewsById(id: String): News? {
-                            return voicesRepository.getNewsById(id)
-                        }
-                    },
+                    voicesEditActions = voicesRepository,
                     leadersList = UserEntity.parseLeadersJson(sharedPrefManager.getCommunityLeaders()),
                     setRepliedNewsIdFn = { sharedPrefManager.setRepliedNewsId(it) }
                 )
