+++ app/src/main/java/org/ole/planet/myplanet/ui/voices/VoicesFragment.kt
@@ -267,17 +267,7 @@ class VoicesFragment : BaseVoicesFragment() {
             },
             onAnimateTyping = VoicesAdapterHelper.createOnAnimateTyping(viewLifecycleOwner.lifecycleScope, dispatcherProvider),
             labelManager = labelManager,
-            voicesEditActions = object : VoicesEditActions {
-                override suspend fun editPost(newsId: String, message: String, imagesToRemove: Set<String>, newImages: List<String>?): News? {
-                    return voicesRepository.editPost(newsId, message, imagesToRemove, newImages)
-                }
-                override suspend fun postReply(message: String, news: News, currentUser: UserEntity, imageList: List<String>?) {
-                    voicesRepository.postReply(message, news, currentUser, imageList)
-                }
-                override suspend fun getNewsById(id: String): News? {
-                    return voicesRepository.getNewsById(id)
-                }
-            },
+            voicesEditActions = voicesRepository,
             leadersList = UserEntity.parseLeadersJson(sharedPrefManager.getCommunityLeaders()),
             setRepliedNewsIdFn = { sharedPrefManager.setRepliedNewsId(it) }
         )
