+++ plan.md
@@ -1,0 +1,19 @@
+1. **Move `VoicesEditActions` to `repository/` and let `VoicesRepository` extend it.**
+   - Rename/move `app/src/main/java/org/ole/planet/myplanet/ui/voices/VoicesEditActions.kt` to `app/src/main/java/org/ole/planet/myplanet/repository/VoicesEditActions.kt` and change its package to `org.ole.planet.myplanet.repository`.
+   - Update `app/src/main/java/org/ole/planet/myplanet/repository/VoicesRepository.kt` to make `interface VoicesRepository : VoicesEditActions`.
+   - Since `VoicesRepository` already has the three methods exactly matching `VoicesEditActions`, we can just remove them from `VoicesRepository` or leave them as overrides. It's better to remove them from `VoicesRepository` to avoid duplication.
+   - Update `VoicesActions.kt` and `VoicesAdapter.kt` to import `org.ole.planet.myplanet.repository.VoicesEditActions` instead of the one in `ui.voices`.
+
+2. **Simplify Instantiation Sites**
+   - In `VoicesFragment.kt`, replace the anonymous object with `voicesEditActions = voicesRepository`.
+   - In `TeamsVoicesFragment.kt`, replace the anonymous object with `voicesEditActions = voicesRepository`.
+   - In `ReplyActivity.kt`, replace the anonymous object with `voicesEditActions = voicesRepository` and remove the unnecessary/incorrect import `import org.ole.planet.myplanet.ui.voices.VoicesEditActions`.
+
+3. **Verify File and Import Updates**
+   - Check files `VoicesFragment.kt`, `TeamsVoicesFragment.kt`, `ReplyActivity.kt`, `VoicesActions.kt`, `VoicesAdapter.kt`, `VoicesRepository.kt`, `VoicesEditActions.kt`.
+
+4. **Run Unit Tests**
+   - Run tests and ensure everything compiles and passes.
+
+5. **Pre-commit step**
+   - Complete pre-commit steps to ensure proper testing, verification, review, and reflection are done.
